pyproject.toml
[project]
name = "ai-essay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "openai>=1.55.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "habanero>=1.2.0",
    "citeproc-py>=0.6.0",
    "semanticscholar>=0.8.0",
    "pytest>=8.0.0",
    "markdown2>=2.4.10",
    "weasyprint>=60.0",
    "python-docx>=1.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
